+++ src/daemon/daemon/navigation.cpp
@@ -16,6 +16,7 @@
 #include "../../core/utils.h"
 #include <PhosphorScreens/Manager.h>
 #include <PhosphorScreens/Swapper.h>
+#include <PhosphorWorkspaces/VirtualDesktopManager.h>
 #include "../../dbus/snapadaptor.h"
 #include "../../dbus/windowtrackingadaptor.h"
 #include <PhosphorIdentity/VirtualScreenId.h>