+++ src/daemon/enginefactory.cpp
@@ -8,6 +8,7 @@
 #include <PhosphorSnapEngine/SnapEngine.h>
 #include <PhosphorEngine/WindowRegistry.h>
 #include <PhosphorPlacement/WindowTrackingService.h>
+#include <PhosphorScreens/Manager.h>
 #include <PhosphorWorkspaces/VirtualDesktopManager.h>
 #include "../core/crosssurfaceresolver.h"
 #include "../core/isettings.h"